refactor: ui 컴포넌트 파일 위치 이동
