+++ src/app/(auth)/layout.tsx
@@ -1,5 +1,5 @@
-import { Button } from "@/components/button";
 import { MobileNavigation } from "@/components/mobile-navigation";
+import { Button } from "@/components/ui/button";
 
 export default function AuthLayout({
   children,
